+++ HISTORY.rst
@@ -11,6 +11,7 @@ UNRELEASED
 * Added class property ``frame_type`` (read-only) to `FitHeader`, `FitCRC`,
   `FitDefinitionMessage` and `FitDataMessage` (``records`` module) to ease and
   speed up type checking
+* Added `FitDataMessage.get_value` method
 
 
 v0.0.1 (2018-07-08)